+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.2.13
+
+- **fix:** tabs height
+- **fix:** DEBUG warning position
+- **feat:** provide/inject
+- **feat:** `row` and `column` components
+- **feat:** removed `lastW` and `lastH`, added `useResolution()`
+
 # 0.2.12
 
 - **fix:** removed duplicate client code